feat!: sqlite (#5)
* SQLite Implementation

!! SQLite does not provide Stored Procedures. Need more work on SQLiteCo object.

* fix!: Namespaces & connectionString checks

- IConnection and DbType renamed,
- Connection removed.

BREAKING CHANGE: new approad of connection instantiation. Doc coming...
