--- Connection/DbType.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace DevHopTools.Connection
-{
-    public enum DbType
-    {
-        MSSQL,
-        MySQL
-    }
-}
