+++ DataAccess/Enums/DatabaseType.cs
@@ -1,0 +1,8 @@
+﻿namespace DevHopTools.DataAccess.Enums
+{
+    public enum DatabaseType
+    {
+        MSSQL,
+        MySQL
+    }
+}
